# API-BACKEND/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fawazlaw Api",
        "description": "Just CMS api",
        "contact": {
            "email": "abderrahman4bouichou@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Retrieve a list of services",
                "operationId": "0f76dda9be6628c47b978d3b7077529a",
                "responses": {
                    "200": {
                        "description": "A list of services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Service"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "description": "Store a newly created resource in storage.",
                "operationId": "ee694c210db03e85b35b350b82c18f1b",
                "requestBody": {
                    "description": "Service data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Service Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Service Description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 100.5
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "example": "service_thumbnail.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Retrieve a specific service",
                "operationId": "ee0c95755e1374f6e92ee6ae1cdddfb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The specified service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update a specific service",
                "description": "Update a specific service.",
                "operationId": "b9a49c8480cb32ec2fa69841690d863e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Service data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Service Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Service Description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 150.75
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "thumbnail": {
                                        "type": "string",
                                        "example": "updated_service_thumbnail.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a specific service",
                "operationId": "cbb6bc64cf61a4812d1c6da5421fd545",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "500": {
                        "description": "Failed to delete service"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/services/map-category": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Map a service to a category",
                "operationId": "e6b295e55d73980f5681454967193d55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "service_id": {
                                        "description": "ID of the service",
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "description": "ID of the category",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service successfully mapped to the category"
                    },
                    "404": {
                        "description": "No mapping found for the provided service and category"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/services/unmap-category": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Unmap a service from a category",
                "operationId": "e45b39d8b89ef547a24e10046e3f721d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "service_id": {
                                        "description": "ID of the service",
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "description": "ID of the category",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service successfully unmapped from the category"
                    },
                    "404": {
                        "description": "No mapping found for the provided service and category"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "title": "Service",
                "description": "Service model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the service",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "The name of the service",
                        "type": "string",
                        "example": "Sample Service Name"
                    },
                    "description": {
                        "description": "The description of the service",
                        "type": "string",
                        "example": "Sample service description"
                    },
                    "price": {
                        "description": "The price of the service",
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "currency": {
                        "description": "The currency used for the price",
                        "type": "string",
                        "example": "SAR"
                    },
                    "thumbnail": {
                        "description": "The URL of the thumbnail image for the service",
                        "type": "string",
                        "example": "https://example.com/thumbnail.jpg"
                    },
                    "created_at": {
                        "description": "The timestamp when the service was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 12:00:00"
                    },
                    "updated_at": {
                        "description": "The timestamp when the service was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 12:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    }
}